requirements.txt
pickle
pandas
streamlit
numpy
sklearn.feature_extraction.text
sklearn.metrics.pairwise
